Got the GET sample working.  This still only has a single FILTER that queries the entire dataset for a particular value.
Added POST query capabilities in the query REST service, including adding a web/postQuerySample.html HTML file that shows how it works.  To be able to query particular fields, you need to use POST.

The names that we query are the URIs of the configuration filenames and NOT the column names in the spreadsheet.  Thus, some mapping may need to happen. The reason for using the URI values is that these are STATIC references to column concepts, in case column names or definitions change.