removed processController's warningsSB as we replace this with commandLineSB.
Updated process.runAllLocally to check for errors. This way the user is alerted when the file can't be opened. I also removed the cmd line information printing from Validation.printMessages. This is now done in process.runAllLocally.

I fixed a bug in the bcidConnector where the cookies were being overwritten when the response didn't return a Set-Cookie header. I fixed another bug in bcidConnector where the access_token was being appended to the url even if the access_token was null. This was preventing bcid from using the session to check that a user is logged in.