Remove field number uniqueness upon request
PK:
I can remove those lines on that XML but John would need to push it. Do you think that's a good change in general, though? Aside from this individual project? (eg, a non-unique field number .. how are you supposed to reference that sample in the future?)
MK:
through the other fields -- like lab no
or occurrence id which is sometimes used
or GUID
we just have to be flexible
PK:
ok
MK:
so in this set: several different collectors and a mix of field and preserved specimens
so for the fieldnum=null often has GUID
ALL have lab no